feat(DTFS2-7445): move view model helpers into subdirectories
