+++ trade-finance-manager-ui/test-helpers/test-data/index.ts
@@ -1,4 +1,3 @@
-export * from './add-payment-view-model';
 export * from './bank-report-period-schedule';
 export * from './bank';
 export * from './tfm-session-user';
