+++ trade-finance-manager-ui/test-helpers/test-data/view-models/cancellation/bank-request-date-view-model.ts
@@ -1,6 +1,6 @@
-import { PRIMARY_NAVIGATION_KEYS } from '../../server/constants';
-import { BankRequestDateViewModel } from '../../server/types/view-models';
-import { aTfmSessionUser } from './tfm-session-user';
+import { PRIMARY_NAVIGATION_KEYS } from '../../../../server/constants';
+import { BankRequestDateViewModel } from '../../../../server/types/view-models';
+import { aTfmSessionUser } from '../../tfm-session-user';
 
 export const aBankRequestDateViewModel = (): BankRequestDateViewModel => ({
   user: aTfmSessionUser(),
