+++ trade-finance-manager-ui/test-helpers/test-data/view-models/cancellation/effective-from-date-view-model.ts
@@ -1,6 +1,6 @@
-import { PRIMARY_NAVIGATION_KEYS } from '../../server/constants';
-import { aTfmSessionUser } from './tfm-session-user';
-import { EffectiveFromDateViewModel } from '../../server/types/view-models';
+import { PRIMARY_NAVIGATION_KEYS } from '../../../../server/constants';
+import { aTfmSessionUser } from '../../tfm-session-user';
+import { EffectiveFromDateViewModel } from '../../../../server/types/view-models';
 
 export const aEffectiveFromDateViewModel = (): EffectiveFromDateViewModel => ({
   user: aTfmSessionUser(),
