+++ trade-finance-manager-ui/test-helpers/test-data/view-models/cancellation/reason-for-cancelling-view-model.ts
@@ -1,6 +1,6 @@
-import { PRIMARY_NAVIGATION_KEYS } from '../../server/constants';
-import { ReasonForCancellingViewModel } from '../../server/types/view-models';
-import { aTfmSessionUser } from './tfm-session-user';
+import { PRIMARY_NAVIGATION_KEYS } from '../../../../server/constants';
+import { ReasonForCancellingViewModel } from '../../../../server/types/view-models';
+import { aTfmSessionUser } from '../../tfm-session-user';
 
 export const aReasonForCancellingViewModel = (): ReasonForCancellingViewModel => ({
   user: aTfmSessionUser(),
