+++ trade-finance-manager-ui/test-helpers/test-data/view-models/index.ts
@@ -1,0 +1,4 @@
+export * from './cancellation';
+export * from './add-payment-view-model';
+export * from './add-to-an-existing-payment-view-model';
+export * from './payment-group-inputs-view-model';
